+++ css/panels.css
@@ -317,7 +317,7 @@
 		color: var(--color-light);
 	}
 	#display_bar .tool, #display_ref_bar > div {
-		width: calc(100% / 8 - 2px);
+		width: calc((100% * 0.125) - 2px);
 		max-width: 52px;
 	}
 	#display_ref_bar > div > label {
