+++ index.html
@@ -27,8 +27,6 @@
 <body spellcheck="false" class="maximized">
 	<script>
 		if (typeof module === 'object') {window.module = module; module = undefined;}//jQuery Fix
-		const isApp = typeof require !== 'undefined';
-
 		
 		if (localStorage.getItem('theme')) {
 			try {
