+++ package.json
@@ -137,6 +137,7 @@
 		"electron-builder": "^24.13.3",
 		"typescript": "^5.8.2",
 		"vite": "^6.2.0",
+		"vite-plugin-conditional-import": "0.1.7",
 		"vite-plugin-electron": "^0.29.0",
 		"vite-plugin-electron-renderer": "^0.14.6",
 		"vite-plugin-pwa": "^0.21.1",
